refactor: 경주 시도 횟수 입력, 유효성 검사 분리 및 static 제거
